+++ src/ESP8266_AT_WebServer.h
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.4.1
+  Version: 1.5.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -36,12 +36,19 @@
   1.3.0   K Hoang      29/05/2021 Add support to Nano_RP2040_Connect, RASPBERRY_PI_PICO using Arduino mbed code
   1.4.0   K Hoang      14/08/2021 Add support to Adafruit nRF52 core v0.22.0+
   1.4.1   K Hoang      08/12/2021 Add Packages_Patches and instructions for BOARD_SIPEED_MAIX_DUINO
+  1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
  *****************************************************************************************************************************/
 
 #ifndef ESP8266_AT_WebServer_h
 #define ESP8266_AT_WebServer_h
 
-#define ESP8266_AT_WEBSERVER_VERSION      "ESP8266_AT_WebServer v1.4.1"
+#define ESP8266_AT_WEBSERVER_VERSION          "ESP8266_AT_WebServer v1.5.0"
+
+#define ESP8266_AT_WEBSERVER_VERSION_MAJOR    1
+#define ESP8266_AT_WEBSERVER_VERSION_MINOR    5
+#define ESP8266_AT_WEBSERVER_VERSION_PATCH    0
+
+#define ESP8266_AT_WEBSERVER_VERSION_INT      1005000
 
 #define USE_NEW_WEBSERVER_VERSION         true
 
@@ -208,6 +215,51 @@ enum HTTPAuthMethod
 #define CONTENT_LENGTH_UNKNOWN  ((size_t) -1)
 #define CONTENT_LENGTH_NOT_SET  ((size_t) -2)
 
+/////////////////////////////////////////////////////////////////////////
+
+#define RETURN_NEWLINE       "\r\n"
+
+#include <Arduino.h>
+
+#if ESP_AT_USE_AVR
+
+  // AVR has no string library
+  typedef String EWString;
+    
+  // Do nothing as this is String
+  #define fromString
+  #define fromEWString 
+
+#else
+
+  #include <string>
+
+  typedef std::string EWString;
+
+  EWString fromString(const String& str)
+  {
+    return str.c_str();
+  }
+
+  EWString fromString(const String&& str)
+  {
+    return str.c_str();
+  }
+
+  String fromEWString(const EWString& str)
+  {
+    return str.c_str();
+  }
+
+  String fromEWString(const EWString&& str)
+  {
+    return str.c_str();
+  }
+
+#endif
+
+/////////////////////////////////////////////////////////////////////////
+
 class ESP8266_AT_WebServer;
 
 typedef struct 
@@ -277,19 +329,20 @@ class ESP8266_AT_WebServer
     }
     #endif
     
-    String arg(String name);            // get request argument value by name
-    String arg(int i);                  // get request argument value by number
-    String argName(int i);              // get request argument name by number
-    int args();                         // get arguments count
-    bool hasArg(String name);           // check if argument exists
+    String arg(const String& name);         // get request argument value by name
+    String arg(int i);                      // get request argument value by number
+    String argName(int i);                  // get request argument name by number
+    
+    int args();                             // get arguments count
+    bool hasArg(const String& name);        // check if argument exists
     void collectHeaders(const char* headerKeys[], const size_t headerKeysCount); // set the request headers to collect
-    String header(String name);         // get request header value by name
-    String header(int i);               // get request header value by number
-    String headerName(int i);           // get request header name by number
-    int headers();                      // get header count
-    bool hasHeader(String name);        // check if header exists
+    String header(const String& name);      // get request header value by name
+    String header(int i);                   // get request header value by number
+    String headerName(int i);               // get request header name by number
+    int headers();                          // get header count
+    bool hasHeader(const String& name);     // check if header exists
 
-    String hostHeader();                // get request host header if available or empty String if not
+    String hostHeader();                    // get request host header if available or empty String if not
 
     // send response to the client
     // code - HTTP response code, can be 200 or 404
@@ -344,8 +397,8 @@ class ESP8266_AT_WebServer
     int  _parseArgumentsPrivate(const String& data, vl::Func<void(String&,String&,const String&,int,int,int,int)> handler);
     bool _parseForm(ESP8266_AT_Client& client, const String& boundary, uint32_t len);
     #else
-    void _parseArguments(String data);
-    bool _parseForm(ESP8266_AT_Client& client, String boundary, uint32_t len);
+    void _parseArguments(const String& data);
+    bool _parseForm(ESP8266_AT_Client& client, const String& boundary, uint32_t len);
     #endif
     
     static String _responseCodeToString(int code);    
@@ -353,6 +406,11 @@ class ESP8266_AT_WebServer
     void _uploadWriteByte(uint8_t b);
     uint8_t _uploadReadByte(ESP8266_AT_Client& client);
     void _prepareHeader(String& response, int code, const char* content_type, size_t contentLength);
+    
+#if !ESP_AT_USE_AVR    
+    void _prepareHeader(EWString& response, int code, const char* content_type, size_t contentLength);
+#endif
+    
     bool _collectHeader(const char* headerName, const char* headerValue);
     
     struct RequestArgument 
@@ -397,7 +455,6 @@ class ESP8266_AT_WebServer
 
     String           _hostHeader;
     bool             _chunked;
-
 };
 
 #include "ESP8266_AT_WebServer-impl.h"