+++ src/utility/RequestHandler.h
@@ -11,7 +11,7 @@
   @file       Esp8266WebServer.h
   @author     Ivan Grokhotkov
 
-  Version: 1.4.1
+  Version: 1.5.0
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -36,6 +36,7 @@
   1.3.0   K Hoang      29/05/2021 Add support to Nano_RP2040_Connect, RASPBERRY_PI_PICO using Arduino mbed code
   1.4.0   K Hoang      14/08/2021 Add support to Adafruit nRF52 core v0.22.0+
   1.4.1   K Hoang      08/12/2021 Add Packages_Patches and instructions for BOARD_SIPEED_MAIX_DUINO
+  1.5.0   K Hoang      19/12/2021 Reduce usage of Arduino String with std::string
  *****************************************************************************************************************************/
 
 #pragma once
@@ -54,7 +55,7 @@ class RequestHandler
 
     virtual ~RequestHandler() { }
 
-    virtual bool canHandle(HTTPMethod method, String uri)
+    virtual bool canHandle(const HTTPMethod& method, const String& uri)
     {
       ESP_AT_UNUSED(method);
       ESP_AT_UNUSED(uri);
@@ -62,14 +63,14 @@ class RequestHandler
       return false;
     }
 
-    virtual bool canUpload(String uri)
+    virtual bool canUpload(const String& uri)
     {
       ESP_AT_UNUSED(uri);
       
       return false;
     }
 
-    virtual bool handle(ESP8266_AT_WebServer& server, HTTPMethod requestMethod, String requestUri)
+    virtual bool handle(ESP8266_AT_WebServer& server, const HTTPMethod& requestMethod, const String& requestUri)
     {
       ESP_AT_UNUSED(server);
       ESP_AT_UNUSED(requestMethod);
@@ -78,7 +79,7 @@ class RequestHandler
       return false;
     }
 
-    virtual void upload(ESP8266_AT_WebServer& server, String requestUri, HTTPUpload& upload) 
+    virtual void upload(ESP8266_AT_WebServer& server, const String& requestUri, const HTTPUpload& upload) 
     {
       ESP_AT_UNUSED(server);
       ESP_AT_UNUSED(requestUri);