+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Releases v1.5.0](#releases-v150)
   * [Releases v1.4.1](#releases-v141)
   * [Releases v1.4.0](#releases-v140)
   * [Releases v1.3.0](#releases-v130)
@@ -35,6 +36,13 @@
 ---
 
 ## Changelog
+
+### Releases v1.5.0
+
+1. Reduce usage of Arduino String with std::string
+2. Optimize library code and examples by using **reference-passing instead of value-passing**.
+3. Update `Packages' Patches`
+
 
 ### Releases v1.4.1